+++ src/app/components/userSettingsEdit/UserSettingsEditComponent.ts
@@ -1,35 +1,23 @@
 import {Input, Component} from 'angular2/core';
-import {Router, RouterLink, CanActivate} from 'angular2/router';
+import {ComponentInstruction, RouterLink, CanActivate} from 'angular2/router';
 import {TranslatePipe} from 'ng2-translate/ng2-translate';
-import {appInjector} from '../../../appInjector';
 
 import {UserSettingsComponent} from '../userSettings/UserSettingsComponent';
 
 import {UserSettingsService} from '../../shared/services/UserSettingsService';
-import {AuthService} from '../../shared/services/AuthService';
+import {AuthService, isLoggedIn} from '../../shared/services/AuthService';
 import {AlertingService} from '../../shared/services/AlertingService';
 
 import {UserSettings} from '../../shared/models/UserSettings';
-import {PointerType, PointerSize, PointerColor, BackgroundColor} from '../../shared/enums/UserSettingsEnums';
 
 @Component({
   directives: [RouterLink, UserSettingsComponent],
   templateUrl: './app/components/userSettingsEdit/userSettingsEdit.html',
   pipes: [TranslatePipe]
 })
-@CanActivate(
-  (nextInstr: any, currInstr: any) => {
-    let injector: any = appInjector();
-    let authService: AuthService = injector.get(AuthService);
-    let router: Router = injector.get(Router);
-    let isLogged = authService.isLogged();
-
-    if (!isLogged) {
-      router.navigate(['/Login']);
-    }
-    return isLogged;
-  }
-)
+@CanActivate((next: ComponentInstruction, previous: ComponentInstruction) => {
+  return isLoggedIn(next, previous);
+})
 export class UserSettingsEditComponent {
   public userName: string;
   public userSettings: UserSettings;
